fix task3.3.4.1
