+++ course3/3.clean_architecture/4.repository/task3.3.4.1/internal/controller/router.go
@@ -5,7 +5,7 @@ import (
 	"github.com/go-chi/chi/middleware"
 )
 
-func NewRouter(controller Controller) *chi.Mux {
+func NewRouter(controller *ImplController) *chi.Mux {
 	router := chi.NewRouter()
 	router.Use(middleware.Logger)
 	router.Use(middleware.Recoverer)
